Fixed tests and services
