+++ src/main/java/com/swave/urnr/chatgpt/config/ChatGPTConfig.java
@@ -1,6 +1,7 @@
 package com.swave.urnr.chatgpt.config;
 
 public class ChatGPTConfig {
+
     public static final String AUTHORIZATION = "Authorization";
     public static final String BEARER = "Bearer ";
     public static final String API_KEY = "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";
@@ -11,4 +12,5 @@ public class ChatGPTConfig {
     public static final String MEDIA_TYPE = "application/json; charset=UTF-8";
     public static final String URL = "https://api.openai.com/v1/chat/completions";
     public static final String ROLE_USER = "user";
+
 }
