+++ src/main/java/com/swave/urnr/project/controller/ProjectController.java
@@ -26,7 +26,6 @@ import java.util.List;
 @Api(tags = "ProjectController")
 public class ProjectController {
 
-
     private final ProjectService projectService;
 
     //0706 : create 생성
