+++ src/main/java/com/swave/urnr/user/service/TokenServiceImpl.java
@@ -1,0 +1,30 @@
+package com.swave.urnr.user.service;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.swave.urnr.user.domain.User;
+import com.swave.urnr.util.oauth.JwtProperties;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.util.Date;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class TokenServiceImpl implements  TokenService{
+
+    @Override
+    public String createToken(User user) {
+
+        String jwtToken = JWT.create()
+                .withSubject(user.getEmail())
+                .withExpiresAt(new Date(System.currentTimeMillis() + JwtProperties.EXPIRATION_TIME))
+                .withClaim("id", user.getId())
+                .withClaim("name", user.getUsername())
+                .sign(Algorithm.HMAC512(JwtProperties.SECRET));
+
+        return jwtToken;
+    }
+}
