+++ src/main/java/com/swave/urnr/user/service/UserInProjectServiceImpl.java
@@ -30,5 +30,6 @@ public class UserInProjectServiceImpl implements UserInProjectService{
                 .message("User drop project")
                 .description(drop +" complete")
                 .build();
+
     }
 }
