+++ src/main/java/com/swave/urnr/util/kafka/KafkaServiceImpl.java
@@ -27,48 +27,18 @@ import java.util.concurrent.ExecutionException;
 @EnableTransactionManagement
 public class KafkaServiceImpl implements  KafkaService{
 
-    // Replace with the topic name you want to produce messages to
+    /*
+    To prevent error before implementation to cloud, temporary  disable Kafka service.
+    It will disappear after fully imported to cloud server during our project, so everything is ok for now.
+     */
+
 
     @Override
     public String createTopic(String topicName) {
-        try {
-            Properties properties = new Properties();
-            properties.put(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
-
-            try (AdminClient adminClient = AdminClient.create(properties)) {
-                NewTopic newTopic = new NewTopic(topicName, 1, (short) 1);
-                adminClient.createTopics(Collections.singleton(newTopic));
-
-               String message = "start";
-                KafkaTemplate<String, String> temp = new KafkaTemplate<>(producerFactory());
-                temp.send(topicName, message);
-
-                return "Topic created successfully: " + topicName;
-            }
-        } catch (Exception e) {
-            return "Failed to create topic: " + e.getMessage();
-        }
+        return "This is test message.";
     }
-
-
- 
-
-
     @Override
     public void produceMessage(NotificationDTO notificationDTO, String topic) {
-        KafkaTemplate<String, String> kafkaTemplate = new KafkaTemplate<>(producerFactory());
-        ObjectMapper objectMapper = new ObjectMapper();
-        log.info(notificationDTO.getContent().toString());
-        log.info(String.valueOf(notificationDTO.getId()));
-        log.info(notificationDTO.getDate().toString());
-        log.info(notificationDTO.getType().toString());
-
-        try {
-            String jsonMessage = objectMapper.writeValueAsString(notificationDTO);
-            kafkaTemplate.send(topic, jsonMessage);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
     }
 
 
@@ -85,16 +55,73 @@ public class KafkaServiceImpl implements  KafkaService{
 
 
 
-    private ProducerFactory<String, String> producerFactory() {
-        Map<String, Object> configs = new HashMap<>();
 
-        configs.put("bootstrap.servers", "localhost:9092");
-        configs.put("key.serializer", "org.apache.kafka.common.serialization.StringSerializer");
-        configs.put("value.serializer", "org.apache.kafka.common.serialization.StringSerializer");
-
-        return new DefaultKafkaProducerFactory<>(configs);
-    }
-
+//    @Override
+//    public String createTopic(String topicName) {
+//        try {
+//            Properties properties = new Properties();
+//            properties.put(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
+//
+//            try (AdminClient adminClient = AdminClient.create(properties)) {
+//                NewTopic newTopic = new NewTopic(topicName, 1, (short) 1);
+//                adminClient.createTopics(Collections.singleton(newTopic));
+//
+//                String message = "start";
+//                KafkaTemplate<String, String> temp = new KafkaTemplate<>(producerFactory());
+//                temp.send(topicName, message);
+//
+//                return "Topic created successfully: " + topicName;
+//            }
+//        } catch (Exception e) {
+//            return "Failed to create topic: " + e.getMessage();
+//        }
+//    }
+//
+//
+//
+//
+//
+//    @Override
+//    public void produceMessage(NotificationDTO notificationDTO, String topic) {
+//        KafkaTemplate<String, String> kafkaTemplate = new KafkaTemplate<>(producerFactory());
+//        ObjectMapper objectMapper = new ObjectMapper();
+//        log.info(notificationDTO.getContent().toString());
+//        log.info(String.valueOf(notificationDTO.getId()));
+//        log.info(notificationDTO.getDate().toString());
+//        log.info(notificationDTO.getType().toString());
+//
+//        try {
+//            String jsonMessage = objectMapper.writeValueAsString(notificationDTO);
+//            kafkaTemplate.send(topic, jsonMessage);
+//        } catch (JsonProcessingException e) {
+//            e.printStackTrace();
+//        }
+//    }
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//    private ProducerFactory<String, String> producerFactory() {
+//        Map<String, Object> configs = new HashMap<>();
+//
+//        configs.put("bootstrap.servers", "localhost:9092");
+//        configs.put("key.serializer", "org.apache.kafka.common.serialization.StringSerializer");
+//        configs.put("value.serializer", "org.apache.kafka.common.serialization.StringSerializer");
+//
+//        return new DefaultKafkaProducerFactory<>(configs);
+//    }
+//
 
 
 }
